requirements.txt
Flask==3.1.1
kafka-python==2.7.7
pymongo==4.13.0
